faster-auto-subtitle/requirements.txt
faster-whisper==1.0.3
tqdm==4.66.4
ffmpeg-python==0.2.0
langcodes==3.3.0
transformers>=4.4,<5
# torch will be installed manually (see below)
numpy>=1.24.2,<=1.26.4
nltk<=3.8.1
sentencepiece<=0.2.0
protobuf<=5.27.2
huggingface_hub<=0.23.5

# ADD BELOW
hangul-romanize
sacremoses
pykakasi
pypinyin

